Refactored ConversationEditor
